new commit

+++ src/main/java/com/bankingapp/root/service/UserService.java
@@ -2,9 +2,11 @@ package com.bankingapp.root.service;
 
 import com.bankingapp.root.common.Constants;
 import com.bankingapp.root.dto.UserDTO;
+import com.bankingapp.root.entity.AccountEntity;
 import com.bankingapp.root.entity.UserEntity;
 import com.bankingapp.root.exception.UserNotFoundException;
 import com.bankingapp.root.mapper.UserDTOEntityMapper;
+import com.bankingapp.root.repository.AccountRepository;
 import com.bankingapp.root.repository.UserRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -16,8 +18,10 @@ import java.util.stream.Collectors;
 @Service
 public class UserService {
     private final UserRepository userRepository;
+    private final AccountRepository accountRepository;
 
-    public UserService(UserRepository userRepository) {
+    public UserService(UserRepository userRepository, AccountRepository accountRepository) {
+        this.accountRepository = accountRepository;
         this.userRepository = userRepository;
     }
 
@@ -27,9 +31,7 @@ public class UserService {
             throw new IllegalArgumentException("User data must not be null.");
         final UserEntity existingUser = userRepository.findById(user.getId())
                 .orElseThrow(() -> new UserNotFoundException("User not found."));
-        existingUser.setUsername(user.getUsername());
         existingUser.setEmail(user.getEmail());
-        existingUser.setPassword(user.getPassword());
         existingUser.setFirstName(user.getFirstName());
         existingUser.setLastName(user.getLastName());
         existingUser.setAddress(user.getAddress());
@@ -61,6 +63,10 @@ public class UserService {
         final UserEntity user = userRepository.findById(id)
                 .orElseThrow(() ->
                         new UserNotFoundException("User not found."));
+        List<AccountEntity> accounts = accountRepository.findByUser_Username(user.getUsername());
+        if (!accounts.isEmpty()) {
+            throw new IllegalStateException("Cannot delete user with existing accounts.");
+        }
         userRepository.delete(user);
         return UserDTOEntityMapper.map(user);
     }
@@ -69,6 +75,10 @@ public class UserService {
     public UserDTO deleteUserByUsername(String username) {
         if (Objects.isNull(username) || username.isEmpty())
             throw new IllegalArgumentException("User data must not be null.");
+        List<AccountEntity> accounts = accountRepository.findByUser_Username(username);
+        if (!accounts.isEmpty()) {
+            throw new IllegalStateException("Cannot delete user with existing accounts.");
+        }
         final UserEntity user = userRepository.findByUsername(username)
                 .orElseThrow(() ->
                         new UserNotFoundException("User not found."));
@@ -83,7 +93,4 @@ public class UserService {
                 .collect(Collectors.toList());
     }
 
-    public UserDTO updateUserRole(Constants.UserRoles userRole) {
-        return null;
-    }
 }
